Added features and a new class
Added a feature for adjusting how long before switching frames on the PlayerImage, added feature which makes the PlayerImage fall due to "gravity", and a method for stopping the gravity/falling, but that method isn't very well fleshed out yet.
Added a class called "Block" which has basic collision detection system on it, for mimicking a floor to fall onto.
I didn't look too far into how collision detection should be done, so for right now it is a bit buggy, and the PlayerImage will get stuck in the floor by varying amounts depending on the speed of gravity and whatnot, especially depending on if you are using Ruffle or Flash Player currently.
I would like to figure out how to implement a better collision detection system when I get the time to.